+++ CHANGELOG.md
@@ -69,6 +69,7 @@ All notable changes to this project will be documented in this file.
 - Values in Dropdown attributes are now sorted alphabetically
 - Values in Dropdown attributes are now searchable
 - Font size in dropdown menus
+- _About_ modal contributors
 
 ## 0.9.14
 ### Added